Review: keep numOctaves as int, fix preset serialization, archive plan
Revert numOctaves from float back to int across all 10 generators —
not modulatable, just a plain SliderInt. Remove int casts and param
registrations added by the standardization. Add numOctaves + baseBright
to PitchSpiralConfig serialization macro. Fix glyph field baseFreq
range comment. Update SPIROL preset. Archive completed plan.

+++ src/effects/arc_strobe.h
@@ -38,7 +38,7 @@ struct ArcStrobeConfig {
 
   // FFT mapping
   float baseFreq = 55.0f;     // Lowest visible frequency in Hz (27.5-440.0)
-  float numOctaves = 5.0f;    // Octave count (1.0-8.0)
+  int numOctaves = 5;         // Octave count (1-8)
   int segmentsPerOctave = 24; // Segments per octave (4-48)
   float gain = 2.0f;          // FFT magnitude amplifier (0.1-10.0)
   float curve = 0.7f;         // Contrast exponent on magnitude (0.1-3.0)
